fix: move files to temp directory to fix github directory capitalization bug
