+++ pom.xml
@@ -9,9 +9,9 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.fatec</groupId>
-	<artifactId>springApi4</artifactId>
+	<artifactId>springapi4</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>springApi4</name>
+	<name>springapi4</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
 		<java.version>17</java.version>
