+++ src/main/java/com/fatec/_springapi4/SpringApi4Application.java
@@ -1,4 +1,4 @@
-package com.fatec.springApi4;
+package com.fatec.springapi4;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
