+++ src/main/java/com/fatec/_springapi4/entity/Partner.java
@@ -1,4 +1,4 @@
-package com.fatec.springApi4.entity;
+package com.fatec.springapi4.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
