+++ src/main/java/com/fatec/_springapi4/entity/PartnerExpertise.java
@@ -1,4 +1,4 @@
-package com.fatec.springApi4.entity;
+package com.fatec.springapi4.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -13,25 +13,29 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
+
 import java.time.LocalDate;
 
 @Getter @Setter @AllArgsConstructor
 @Entity
-@Table(name = "Partner_Track", uniqueConstraints = { @UniqueConstraint(columnNames = { "pt_id", "tk_id" }) })
-public class PartnerTrack {
+@Table(name = "Partner_Expertise", uniqueConstraints = { @UniqueConstraint(columnNames = { "pt_id", "ex_id" }) })
+public class PartnerExpertise {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "pt_tk_id")
+    @Column(name = "pt_ex_id")
     private Long id;
     
     @ManyToOne
-    @JoinColumn(unique=true, name = "pt_id", referencedColumnName = "pt_id")
+    @JoinColumn(name = "pt_id", referencedColumnName = "pt_id")
     private Partner partnerId;
     
     @ManyToOne
-    @JoinColumn(name = "tk_id", referencedColumnName = "tk_id")
-    private Track trackId;
+    @JoinColumn(name = "ex_id", referencedColumnName = "ex_id")
+    private Expertise expertiseId;
     
-    @Column(name = "pt_tk_insert_date")
+    @Column(name = "pt_ex_insert_date")
     private LocalDate insertDate;   
+    
+    @Column(name = "pt_ex_complete_date")
+    private LocalDate completeDate;   
 }
