+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=springApi4
+spring.application.name=springapi4
