+++ src/test/java/com/fatec/springAPI4/SpringApi4ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.fatec.springApi4;
+package com.fatec.springapi4;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
